+++ packages/flutter_monaco_editor/lib/src/models/models.dart
@@ -7,3 +7,4 @@ export 'monaco_position.dart';
 export 'monaco_range.dart';
 export 'monaco_scroll_event.dart';
 export 'monaco_selection.dart';
+export 'monaco_theme.dart';